# openapi.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/jamesvincentsiauw/tubes_lasti/1.0.0"
  }, {
    "url" : "http://178.128.104.74/pengelolaanjalurseleksipmb/"
  } ],
  "info" : {
    "description" : "API Penerimaan Mahasiswa Baru",
    "version" : "1.0.0",
    "title" : "Penerimaan Mahasiswa Baru Kelompok 1",
    "contact" : {
      "email" : "jamesvincentsiauw@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags" : [ {
    "name" : "admins",
    "description" : "Khusus admin sistem"
  }, {
    "name" : "students",
    "description" : "Khusus pendaftar"
  }, {
    "name" : "auth",
    "description" : "Untuk autentikasi user"
  } ],
  "paths" : {
    "/student/requirements" : {
      "get" : {
        "tags" : [ "students" ],
        "summary" : "get required file",
        "responses" : {
          "200" : {
            "description" : "menampilkan berkas yang dibutuhkan",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/RequiredFiles"
                      },
                      "example" : [ {
                        "jalur" : "Seleksi Mandiri",
                        "id" : "SM5627",
                        "files" : "Ijazah SMA"
                      }, {
                        "jalur" : "Seleksi Mandiri",
                        "id" : "SM3671",
                        "files" : "Fotokopi KTP"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "post" : {
        "tags" : [ "students" ],
        "summary" : "upload required files",
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/Files"
              }
            }
          }
        },
        "parameters" : [ {
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Data berhasil diupload",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "description" : {
                      "type" : "string",
                      "format" : "desc",
                      "example" : "Data Uploaded"
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/Files"
                      },
                      "example" : [ {
                        "idpendaftar" : "RIWV2J8PFPKJTG1AX3QREY5513C5Z9",
                        "jalur" : "Seleksi Mandiri",
                        "idfiles" : "SMJ02",
                        "filepath" : "/uploads/requirements/02.pdf",
                        "uploadeddate" : "23-12-2019",
                        "verified" : false
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/admin/jadwal/{id}" : {
      "put" : {
        "tags" : [ "admins" ],
        "summary" : "update jadwal tahapan seleksi",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/TahapanSeleksi"
              }
            }
          }
        },
        "responses" : {
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      },
      "delete" : {
        "tags" : [ "admins" ],
        "summary" : "menghapus jadwal tahapan seleksi",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Data deleted"
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/admin/jadwal" : {
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "menambahkan jadwal tahapan seleksi",
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/TahapanSeleksi"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Data berhasil dimasukan",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "description" : {
                      "type" : "string",
                      "format" : "desc",
                      "example" : "Data Inserted"
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/TahapanSeleksi"
                      },
                      "example" : [ {
                        "id" : "69FZFFCBLAL551TU4ZE5",
                        "tanggalMulai" : "15 Okktober 2019",
                        "tanggalAkhir" : "1 Desember 2019",
                        "kegiatan" : "Batas akhir pendaftaran peserta seleksi"
                      }, {
                        "id" : "OVYL7KF5934JZY0NPOMF",
                        "tanggalMulai" : "2 Desember 2019",
                        "tanggalAkhir" : "5 Desember 2019",
                        "kegiatan" : "Pengumuman peserta lolos seleksi"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad request"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/student/jadwal" : {
      "get" : {
        "tags" : [ "students" ],
        "summary" : "melihat jadwal tahapan seleksi",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "jalurn" : {
                      "type" : "string",
                      "format" : "jalur",
                      "example" : "Seleksi Mandiri"
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/TahapanSeleksi"
                      },
                      "example" : [ {
                        "id" : "69FZFFCBLAL551TU4ZE5",
                        "tanggalMulai" : "15 Okktober 2019",
                        "tanggalAkhir" : "1 Desember 2019",
                        "kegiatan" : "Batas akhir pendaftaran peserta seleksi"
                      }, {
                        "id" : "OVYL7KF5934JZY0NPOMF",
                        "tanggalMulai" : "2 Desember 2019",
                        "tanggalAkhir" : "5 Desember 2019",
                        "kegiatan" : "Pengumuman peserta lolos seleksi"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/student/files" : {
      "get" : {
        "tags" : [ "students" ],
        "summary" : "menampilkan daftar berkas yang sudah di upload oleh peserta",
        "parameters" : [ {
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "berkas yang sudah diupload",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "identity" : {
                      "type" : "object",
                      "properties" : {
                        "idpendaftar" : {
                          "type" : "string",
                          "format" : "id",
                          "example" : "RIWV2J8PFPKJTG1AX3QREY5513C5Z9"
                        },
                        "jalur" : {
                          "type" : "string",
                          "format" : "jalur",
                          "example" : "Seleksi Mandiri"
                        }
                      }
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/Files"
                      },
                      "example" : [ {
                        "idfiles" : "SM5627",
                        "filepath" : "uploads/requirements/02.pdf",
                        "uploadedAt" : "7 Oktober 2019",
                        "verified" : true
                      }, {
                        "idfiles" : "SM3671",
                        "filepath" : "uploads/requirements/03.pdf",
                        "uploadedAt" : "11 Oktober 2019",
                        "verified" : false
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/auth/register" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "register new user",
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "user registered",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "description" : {
                      "type" : "string",
                      "format" : "desc",
                      "example" : "User Registered"
                    },
                    "results" : {
                      "type" : "array",
                      "items" : {
                        "$ref" : "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/auth/login" : {
      "post" : {
        "tags" : [ "auth" ],
        "summary" : "login system",
        "requestBody" : {
          "content" : {
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "type" : "object",
                "required" : [ "username", "password" ],
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "format" : "username",
                    "example" : "jamesvincentsiauw"
                  },
                  "password" : {
                    "type" : "string",
                    "format" : "password",
                    "example" : "*******"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Login Sukses",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "sessionID" : {
                      "type" : "string",
                      "format" : "sessionID",
                      "example" : "9121882919SS810"
                    },
                    "results" : {
                      "type" : "string",
                      "example" : "Berhasil Login Vincent"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    },
    "/auth/logout" : {
      "post" : {
        "tags" : [ "auth" ],
        "description" : "logout system",
        "responses" : {
          "200" : {
            "description" : "Logout Sukses",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "status" : {
                      "type" : "integer",
                      "format" : "status",
                      "example" : 200
                    },
                    "results" : {
                      "type" : "string",
                      "example" : "Berhasil Logout"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          },
          "500" : {
            "description" : "internal server error"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "type" : "object",
        "required" : [ "nama", "fakultas", "alamat", "ttl", "tingkat", "jalur", "username", "password", "email" ],
        "properties" : {
          "nama" : {
            "type" : "string",
            "format" : "name",
            "example" : "Vincent"
          },
          "alamat" : {
            "type" : "string",
            "format" : "address",
            "example" : "Jalan Bandung no 780"
          },
          "fakultas" : {
            "type" : "string",
            "format" : "faculty",
            "example" : "STEI"
          },
          "ttl" : {
            "type" : "string",
            "format" : "ttl",
            "example" : "7 Januari 1998"
          },
          "tingkat" : {
            "type" : "string",
            "format" : "tingkat",
            "example" : "S1"
          },
          "jalur" : {
            "type" : "string",
            "format" : "jalur",
            "example" : "Seleksi Mandiri"
          },
          "username" : {
            "type" : "string",
            "format" : "username",
            "example" : "jamesvincentsiauw"
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "example" : "alamak@ahaydeu.com"
          },
          "password" : {
            "type" : "string",
            "format" : "password",
            "example" : "********"
          }
        }
      },
      "RequiredFiles" : {
        "type" : "object",
        "properties" : {
          "jalur" : {
            "type" : "string",
            "format" : "jalur"
          },
          "id" : {
            "type" : "string",
            "format" : "id"
          },
          "files" : {
            "type" : "string",
            "format" : "file"
          }
        }
      },
      "TahapanSeleksi" : {
        "type" : "object",
        "required" : [ "tanggalMulai", "tanggalAkhir", "kegiatan" ],
        "properties" : {
          "tanggalMulai" : {
            "type" : "string",
            "format" : "startDate"
          },
          "tanggalAkhir" : {
            "type" : "string",
            "format" : "endDate"
          },
          "kegiatan" : {
            "type" : "string",
            "format" : "activity"
          }
        }
      },
      "Files" : {
        "type" : "object",
        "required" : [ "idpendaftar", "jalur", "idfiles", "file", "uploadeddate" ],
        "properties" : {
          "idpendaftar" : {
            "type" : "string",
            "format" : "id"
          },
          "jalur" : {
            "type" : "string",
            "format" : "jalur"
          },
          "idfiles" : {
            "type" : "string",
            "format" : "id"
          },
          "file" : {
            "type" : "object",
            "format" : "file"
          },
          "uploadeddate" : {
            "type" : "string",
            "format" : "uploadedAt"
          }
        }
      }
    }
  }
}